checkpoint: MessageList media utils extraction (2347→2057)
